linkedin/test-butler#12 pull request comments improvements
