+++ test-butler-library/src/main/java/com/linkedin/android/testbutler/TestButler.java
@@ -158,6 +158,22 @@ public class TestButler {
     }
 
     /**
+     * Enable/disable the Wifi connection on the emulator
+     *
+     * @param enabled true if wifi should be enabled, false otherwise
+     */
+    public static void setGsmState(boolean enabled) {
+        verifyApiReady();
+        try {
+            if (!butlerApi.setGsmState(enabled)) {
+                throw new IllegalStateException("Failed to set gsm state!");
+            }
+        } catch (RemoteException e) {
+            throw new IllegalStateException("Failed to communicate with ButlerService", e);
+        }
+    }
+
+    /**
      * Change the location services mode on the emulator
      *
      * @param locationMode one of the {@link LocationMode} IntDef values
